added download settings

+++ convert.py
@@ -1,13 +1,13 @@
 import os
 
 
-def convert_to_wav(path):
-    if os.path.isdir(path):
+def convert_to(path, exp_type: str):
+    if os.path.isdir(path) and exp_type != ".ogg":
         src_list = os.listdir(path)
 
         for audio in src_list:
             try:
-                os.system(f"ffmpeg.exe -i ./{path}/{audio} ./{path}/{audio[:-4]}.wav")
+                os.system(f"ffmpeg.exe -i ./{path}/{audio} ./{path}/{audio[:-4]}{exp_type}")
                 os.remove(f"./{path}/{audio}")
                 print(f"./{path}/{audio} 파일 삭제")
             except Exception as e:
@@ -18,4 +18,4 @@ if __name__ == "__main__":
     character = input("원하는 블루아카 캐릭터를 영문으로 입력하세요(예시: Azusa) : ")
     path = f"./{character.replace(' ', '_')}/"
 
-    convert_to_wav(path)
+    convert_to(path, ".wav")
